+++ UpgradeWorld/Operations/locations/RemoveLocations.cs
@@ -2,29 +2,37 @@ using System.Collections.Generic;
 using System.Linq;
 namespace UpgradeWorld;
 ///<summary>Destroys given locations.</summary>
-public class RemoveLocations : ExecutedOperation {
+public class RemoveLocations : ExecutedOperation
+{
   readonly HashSet<string> Ids;
   readonly FiltererParameters Args;
-  public RemoveLocations(Terminal context, IEnumerable<string> ids, FiltererParameters args) : base(context, args.Start) {
-    Args = new(args) {
+  public RemoveLocations(Terminal context, IEnumerable<string> ids, FiltererParameters args) : base(context, args.Start)
+  {
+    Args = new(args)
+    {
       TargetZones = TargetZones.All
     };
     Ids = ids.ToHashSet();
   }
 
-  private int RemoveSpawned() {
+  private int RemoveSpawned()
+  {
     var zs = ZoneSystem.instance;
     var zdos = ZDOMan.instance.m_objectsByID.Values.Where(zdo => LocationProxyHash == zdo.GetPrefab());
     zdos = Args.FilterZdos(zdos);
     var removed = 0;
-    foreach (var zdo in zdos) {
+    foreach (var zdo in zdos)
+    {
       if (!Args.Roll()) continue;
       var zone = zs.GetZone(zdo.GetPosition());
       var name = "???";
-      if (zs.m_locationsByHash.TryGetValue(zdo.GetInt(LocationHash), out var location)) {
+      if (zs.m_locationsByHash.TryGetValue(zdo.GetInt(LocationHash), out var location))
+      {
         name = location.m_prefabName;
         if (Ids.Count > 0 && !Ids.Contains(name)) continue;
-        Helper.ClearZDOsWithinDistance(zone, zdo.GetPosition(), Args.ObjectReset != 0 ? location.m_location.m_exteriorRadius : Args.ObjectReset);
+        var clearRadius = location.m_location.m_exteriorRadius;
+        if (Args.ObjectReset.HasValue) clearRadius = Args.ObjectReset.Value;
+        Helper.ClearZDOsWithinDistance(zone, zdo.GetPosition(), clearRadius);
       }
       Helper.RemoveZDO(zdo);
       removed++;
@@ -33,13 +41,16 @@ public class RemoveLocations : ExecutedOperation {
         Print($"Location {name} removed at {zone}.");
     }
     var toRemove = zs.m_locationInstances.Where(kvp => Args.FilterPosition(kvp.Value.m_position)).ToList();
-    foreach (var kvp in toRemove) {
+    foreach (var kvp in toRemove)
+    {
       if (!Args.Roll()) continue;
       var zone = kvp.Key;
       var location = kvp.Value.m_location;
       var name = location.m_prefabName;
       if (Ids.Count > 0 && !Ids.Contains(name)) continue;
-      Helper.ClearZDOsWithinDistance(zone, kvp.Value.m_position, Args.ObjectReset != 0 ? location.m_location.m_exteriorRadius : Args.ObjectReset);
+      var clearRadius = location.m_location.m_exteriorRadius;
+      if (Args.ObjectReset.HasValue) clearRadius = Args.ObjectReset.Value;
+      Helper.ClearZDOsWithinDistance(zone, kvp.Value.m_position, clearRadius);
       ResetTerrain.ResetRadius = Args.TerrainReset == 0f ? location.m_exteriorRadius : Args.TerrainReset;
       ResetTerrain.Execute(kvp.Value.m_position);
       removed++;
@@ -51,7 +62,8 @@ public class RemoveLocations : ExecutedOperation {
     return removed;
 
   }
-  private int RemoveNotSpawned() {
+  private int RemoveNotSpawned()
+  {
     var zs = ZoneSystem.instance;
     var filterers = FiltererFactory.Create(Args);
     if (Ids.Count > 0)
@@ -59,7 +71,8 @@ public class RemoveLocations : ExecutedOperation {
     List<string> messages = new();
     var notSpawnedZones = filterers.Aggregate(Zones.GetZones(TargetZones.All), (zones, filterer) => filterer.FilterZones(zones, ref messages));
     var removed = 0;
-    foreach (var zone in notSpawnedZones) {
+    foreach (var zone in notSpawnedZones)
+    {
       if (!Args.Roll()) continue;
       if (!zs.m_locationInstances.TryGetValue(zone, out var location)) continue;
       removed++;
@@ -70,13 +83,15 @@ public class RemoveLocations : ExecutedOperation {
     return removed;
   }
 
-  protected override bool OnExecute() {
+  protected override bool OnExecute()
+  {
     var removed = RemoveSpawned() + RemoveNotSpawned();
     Print($"Removed {removed} locations.");
     return true;
   }
 
-  protected override string OnInit() {
+  protected override string OnInit()
+  {
     return Args.Print($"Remove locations{Helper.LocationIdString(Ids)} from");
   }
 }