+++ UpgradeWorld/Settings.cs
@@ -38,10 +38,9 @@ public static class Settings
   public static int DestroysPerUpdate => configDestroysPerUpdate.Value;
   public static ConfigEntry<string> configTimeBasedDataNames;
   public static IEnumerable<string> TimeBasedDataNames => configTimeBasedDataNames.Value.Split(',').Select(name => name.Trim());
-  public static ConfigEntry<string> configZoneControlId;
-  public static string ZoneControlId => configZoneControlId.Value.Trim();
-  public static ConfigEntry<string> configTerrainCompilerId;
-  public static string TerrainCompilerId => configTerrainCompilerId.Value.Trim();
+  public static int ZoneControlHash = "_ZoneCtrl".GetStableHashCode();
+
+  public static int TerrainCompilerHash = "_TerrainCompiler".GetStableHashCode();
   public static ConfigEntry<string> configRootUsers;
   public static HashSet<string> RootUsers => configRootUsers.Value.Split(',').Select(s => s.Trim()).Where(s => s != "").ToHashSet();
 
@@ -62,7 +61,5 @@ public static class Settings
 
     configDestroysPerUpdate = config.Bind("2. Destroying", "Operations per update", 100, "How many zones are destroyed per Unity update.");
     configTimeBasedDataNames = config.Bind("3. Change time/day", "Time based data names", "spawntime,lastTime,SpawnTime,StartTime,alive_time,spawn_time,picked_time,plantTime,pregnant,TameLastFeeding", "Names of the data values that should be updated with the new time. Changing these is NOT recommended.");
-    configZoneControlId = config.Bind("4. Other", "Zone control name", "_ZoneCtrl", "Name of the zone control entity which controls the enemy spawning. Changing these is NOT recommended.");
-    configTerrainCompilerId = config.Bind("4. Other", "Terrain compiler name", "_TerrainCompiler", "Name of the terrain compiler keeping track of terrain edits. Changing these is NOT recommended.");
   }
 }